Make Audiences::Scim a module
