--- audiences/app/models/audiences/safe_object.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Audiences
-  SafeObject = Struct.new(:id, :displayName, :photos) do
-    def initialize(attrs)
-      super attrs["id"], attrs["displayName"], attrs["photos"]
-    end
-  end
-end
