+++ audiences/spec/dummy/config/initializers/scim_proxy.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Audiences.scim = Audiences::Scim.new(
+Audiences.scim = Audiences::Scim::Client.new(
   uri: ENV.fetch("SCIM_V2_API", "http://example.com/scim/v2/"),
   headers: { "Authorization" => "Bearer 123456789" }
 )
